+++ test/ContentScript/unitTests.test.ts
@@ -169,10 +169,10 @@ test('Report page forms', () => {
   // Then
   expect(mockFn.mock.calls.length).toBe(2);
   expect(mockFn.mock.calls[0][0].toNonTimestampString()).toBe(
-    '{"type":"nodeAdded","tagName":"FORM","id":"form1","nodeName":"FORM","url":"http://localhost/","text":"Content1"}'
+    '{"type":"nodeAdded","tagName":"FORM","id":"form1","nodeName":"FORM","url":"http://localhost/","text":"Content1","formId":-1}'
   );
   expect(mockFn.mock.calls[1][0].toNonTimestampString()).toBe(
-    '{"type":"nodeAdded","tagName":"FORM","id":"form2","nodeName":"FORM","url":"http://localhost/","text":"Content2"}'
+    '{"type":"nodeAdded","tagName":"FORM","id":"form2","nodeName":"FORM","url":"http://localhost/","text":"Content2","formId":-1}'
   );
 });
 
@@ -263,7 +263,7 @@ test('Reported page loaded', () => {
     '{"type":"nodeAdded","tagName":"AREA","id":"","nodeName":"AREA","url":"http://localhost/","href":"https://www.example.com/1","text":""}'
   );
   expect(mockFn.mock.calls[2][0].toNonTimestampString()).toBe(
-    '{"type":"nodeAdded","tagName":"FORM","id":"form1","nodeName":"FORM","url":"http://localhost/","text":"FormContent"}'
+    '{"type":"nodeAdded","tagName":"FORM","id":"form1","nodeName":"FORM","url":"http://localhost/","text":"FormContent","formId":-1}'
   );
   expect(mockFn.mock.calls[3][0].toNonTimestampString()).toBe(
     '{"type":"nodeAdded","tagName":"INPUT","id":"input1","nodeName":"INPUT","url":"http://localhost/","text":"default","tagType":"text"}'