+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Changed
+- Report index of the forms.
+
 ## 0.0.9 - 2024-11-28
 
 ### Added